extract parse endpoint from IConfiguration method to a extension method
